Remove unused / unecessary clutter
- Removed MOM6_compute_transport because it is not useful
- Removed helper functions for MOM6_compute_transport that no longer serve a purpose
- Removed MOM6_normal_transport because it is an arbitrary special case of MOM6_convergent_transport
- Removed anything related to offset_center because MOM6_UVpoints and MOM_convergent_transport now correctly handle both symmetric and non-symmetric grids--all it takes is setting the keyword argument from symmetric=True to symmetric=False

Note: Subsetting the grid such that it does not preserve the symmetric/non-symmetric grid characteristics is no longer supported by these changes. It is possbible to add this functionality back in, but it should include extensive testing or examples if that is going to be the case. It was not currently being used in either of the example notebooks, and I could not vouch for its accuracy / correctness.