Ignore duplicate edge points for nonsymmetric grid paths!
Because the `infer_grid_path` function finds the shortest path
in index-space, the algorithm incorrectly thinks that the duplicate
points at periodic boundaries of a non-symmetric grid are different,
and it effectively double counts them! This gives spurious results
for transports along sections that cross the periodic boundaries.

This has been fixed by requiring the Boolean argument `symmetric` for `create_section`
and `create_section_composite` (the latter passes the argument to the former). If False,
the input grid is made symmetric by removing the first grid row and column. Because this
then gives indices that are offset by one, the indices output by `infer_grid_path_form_geo`
are also incremented before returning.