Reproduced `convergent_transport` with new (more efficient xarray-enabled) algorithm
Follows the basic structure from the earlier prototype, but exactly reproduces the for-loop
version of the algorithm! I think the bug in https://github.com/raphaeldussin/sectionate/blob/master/sectionate/transports_C.py#L275-L280 that was previously preventing the two versions of the algorithm from converging, is the treatment of the NaNs. In the old algorithm, NaNs in the U field do not affect NaNs in the Vfield. However, the way masks are applied to the velocity components before adding them in the prior implementation of the new algorithm does not mask out the NaNs, so then adding a non-zero transport value to the supposedly masked NaN in the other component just results in a NaN. I fixed this by simply adding a call to the `.fill_na(0.)` method before adding the velocity components.

On the OSNAP example, the xarray version of the `convergent_transport` function is about 5 times faster than the for-loop version. I suspect this speed-up will also scale non-linearly so will be a massive speed-up for big-data applications!