Corrected convention for "symmetric" grids (I had it backwards)
I got confused with how MOM6 defines a symmetric grid and thought that it
refered to when the vorticity and tracer grids have the same shape, but it
is in fact the opposite. I went through all of the source code and corrected
all of the relevant sign errors (in grid indexing) and logical operators.

I also updated all of the example notebooks accordingly. This bug fix does not change any
quantitative results but now the convention for "symmetric" matches the MOM6 docs.