Restructure API to make better use of metadata embedded in xgcm.Grid
- Added `gridutils.py` which contains some helpful functions that parse
xgcm.Grid instances to get needed metadata about grid coordinates and
their geometries.

- Added `grid_section` as a simple wrapper for `create_section_composite`
that passes the kwargs consistent with the grid's metadata

- Removed `closed` argument to `create_section_composite`, since this was
confusing and redundant. A user can simple pass a section that has the same
start and end points.

- Changed support `periodic` behavior to be periodic in `X` only. It was
previously assuming periodic in `Y` in some places in the code, but not
everywhere, which gave inconsistent results.

- Added missing factor of 1/2 to distance formula.

- Improved formatting in many places, mostly just by indenting

- Fixed wrapping of indices in `uvcoords_from_uvindices` in case
where `geolon_u` and `geolon_v` are not directly provided and
they must be infered from `geolon` and `geolon_c`.

- The dataset returned by `convergent_transport` now includes
several dependent coordinates with useful information, e.g. about
whether each velocity face is 'U' or 'V', and whether or not the
sign of U,V needs to be flipped to match the "convergent" orientation.

- Includes new `is_mask_inside` function that can be used to verify
the orientation of a boundary. (This is particularly useful when
combined with the `regionate` package, https://github.com/hdrake/regionate).