[WIP] Attempt at using only (gridlon, gridlat) information for grid paths
The goal is to find the shortest path of consecutive neighbors that gets
from the starting coordinates to the end coordinates. Ideally this series
of velocity faces connecting grid corners would closely follow the
geodesic arcs, which are the shortest non grid-conforming paths.

In practice, however, I am having difficult devising an algorithm that
consistently picks the next neighbor in the sequence.