Scrapped index-based grid path algo for spherical trigonometric one
There were three fundamental problems with the old index-based
algorithm in the core function `infer_grid_path` (formerly
`infer_broken_line`):
  1. It assumed a straight line in index space points in the direction
of the shortest path, which is not true for more general grid topologies
(e.g. the tripolar seam in the Arctic region of OM4).
  2. It searches the shortest path (straight line) in (lon,lat)
mercator coordinates, which are not actually the shortest path on the
surface of a sphere (i.e. great circle arcs, or geodesic paths).
  3. It does not take into account the anisotropy of grid cell widths,
since it only compares distances in index space.

The new complete overhaul of the core `infer_grid_path` function now
addresses all three of these points by using robust
spherical trigonometric identities to identify the list of consecutive grid
indices which most closely follows the shortest path on the sphere.

The algorithm has three simple steps:
1) identify the valid neighbors of each point in index space, including
a treatment of edge cases, such as periodicity in the ('X') grid dimension
and the seam at the upper `j=ny-1` index of OM4's tripolar grid.
2) check which of these valid neighbors gets us closer to the target end point,
relative to the current point.
3) check which of the remaining sub-set of valid points has the smallest angle
the shortest path between the start and end points.

The algorithm is conceptually the same as the previously improved one, only no
longer bakes in assumptions about the model grid topology and generalizes to
distances between points defined by spherical coordinates (lon,lat) on the
surface of a sphere.

This commit is breaking in the sense that the results will now be quite different,
because the identified shortest paths are very different as they now closely
follow shortest paths on the surface of a sphere. Answers to some unit tests
have been updated accordingly. (In the limiting cases of purely zonal or purely
meridional sections, the results are unchanged).

A major corrolary of this reformulation is that `sectionate` now supports sections
in the high Arctic region of OM4 where tripolar seams move singularities on land,
whereas it previous failed to find the shortest paths (or, perhap worse, produced
incorrect paths distorted by the tripolar folds).