Use one-sided interpolation if one of two adjacent cells is NaN
- For non-native coordinate frames, the regridding of transports and tracers can become
inconsistent because horizontal transports are at constant natv levels, which are not
necessarily the same as constant levels of other coordinate. In particular, near steep slopes
such as continental slopes, `extract_tracer` was giving NaNs for cell faces that have non-zero
transports. Importantly, this bug can result in O(1) errors in lambda-binned transports (as in
Water Mass Budget calculations of `xwmb`).
- The new behavior fixes this by defaulting to whichever one-sided tracer value is not NaN in
case one of the two is. Verified this works as expected by plotting the tracer and transport sections
and by verifying the vertical integral of transports in the original coordinate is reproduced in the
transformed target lambda coordinate.