Update xgcm's periodic BC API
- Changes to `xgcm`'s API for setting axes' boundary conditions broke
the way we were created xgcm.Grid instances with `periodic=['X']`, which
was incorrectly propagated periodic BC to all axes, which causes problems
with other packages when they try to transform in 'Z'. Instead, we now
use the `boundary` argument, e.g. with `boundary={'X':'periodic', 'Z':'extend'}`.
We also now access this metadata from the Grid instance with:
`grid.axes["X"]._boundary is "periodic"`

- Minor improvement/bugfix to the `infer_grid_path` algorithm to better deal with cases
where there are two plausible options that get you closer to the your target.

- Changed single quotes to double quotes

- Renamed `4_edge_case_behavior.ipynb` -> `4_sections_on_global_tripolar_grid.ipynb`
and added an example of how to deal with the tripolar grid fold issues

- Updated/re-ran all notebooks