Change default topology to "latlon", fix tests, and re-run notebooks
- Changed the default topology from "cartesian" to "latlon". Since most
of the tests and current applications of the package are to "latlon" coordinates,
and since this is the most common and basic ocean model topology, we changed this
default. Eventually this keyword argument will be completely deprecated because the
grid topology will be inferred from the `boundary` and `face_connections` metadata
input to the `xgcm.Grid` instance that is passed to sectionate. See the discussion of
this here: https://github.com/xgcm/xgcm/issues/194.

- Tests were no longer consistent with the upgraded `xgcm` dependency that uses the
`boundary` kwarg to `xgcm.Grid` instead of the (currently erroneous and soon-to-be-deprecated
`periodic` kwarg).