Use average of angle from both directions to guarantee path symmetry
With the previous version of the algorithm, the path chosen differs slightly dependong on which end point you start with. This path-dependence is problematic for both reproducibility but also for internal consistency within a given project. For example, if two regions share a boundary, then sectionate may give different discrete boundaries for the two separate regions if their shared boundary section is traveling in different directions.