Use np.nanargmin to find start/end point when there are NaNs in coordinates
In some OM5dev runs, the coordinates `geolon_c` and `geolat_c` for some reason have values replaced by NaNs,
sometimes at the edges of the domain and other times on land. While this is probably undesirable for many reasons,
it should not affect the behavior of `sectionate` *away from* these NaN values.

Thanks to @33de6maggie for pointing out that any NaNs in the coordinates were perplexingly causing
sectionate to give errors, even if computing sections far from the NaN regions! I traced the problem
to the `dist.argmin()` call that is used to find the indices of the grid corners closest to the target
section segment's start and end points. The problem is resolved by replacing this np.array method call
with a call to the more robust `np.nanargmin` function that skips NaN values when finding the minimum.