Removes unnecessary dependency on pip: xoverturning
Relying on xoverturning, which is installed via pip based on
a github link, introduces problems. The dependencies for xoverturning
are incompletely listed in the setup files, and cmip_basins and
regionmask has to be separately specified. These are not even used
in our sectionate examples, so I propose we remove them.

The streamfunctions are much more transparently computed by summing
the sectionate transports. xoverturning was more useful before
because it was used to transform z-coordinate output to sigma2 coordinates,
but this functionality is better left to the xoverturning docs.